OI Shopping List: Merge shoppinglist_filters branch to trunk.
Added a new preference for OI Shopping List called "filters." After it is
turned on, the spinner to select lists is replaced by three buttons: List,
Tags, and Stores. Picking any of those buttons summons a PopupMenu. The Lists
menu replaces the functionality of the old spinner. The Tags and Stores
buttons allow you to select (Unfiltered) or the name of a tag or store.
If you select the name of a tag or store, the label of the button is replaced
with the tag or store you selected, and the list is filtered to only show
matching items. For the Stores filter, the price is changed from the cheapest
price at any store to the price at the selected store.

git-svn-id: https://openintents.googlecode.com/svn/trunk@4094 72b678ce-9140-0410-bee8-679b907dd61a