More layout tweaks.
Tries to position price based on existence of tags.

Wanted to also allow name to extend into last column if price is
missing or on tags line, but that causes strange behavior, so
commented for now.

Still only for layout_shopping_item_small, and currently the long-press
menu is not working.