ShoppingItemsView refactoring via mItemRowState.
newView makes a class containing all the subordinate views of the
current row, binds their callbacks, and sets that class as the tag
of all the views. bindView sets the cursor position within that class,
where the item click listeners look for it.

Should be faster due to not needing findViewById calls when rebinding
a set of views from one cursor position to another. Also, less duplicated
logic, using two shared classes for the slight variations of
ItemClickListeners.

Main purpose, though, was to facilitate access to price while computing
the subfields of item name -- need to know how wide to make a transparent
image to overlay the price. Fixing the size of that image should be the
next step.