Extra functionality in widget
I think it is too easy to hit an item and mark it as 'bought' by
accident.  And after it happened, it is difficult to know which item
it was (if you even noticed that this happened).
Therefore I think it is better to display the 'bought' items grayed out,
and have an extra button in the widget to clean the list.
I chose the default 'delete' menu icon, and not the icon as used in
the app, because that one is really difficult to see on the light
widget background.

Also aligned the icons and title text to the top of the widget, because
on a tablet a widget with a large height had the buttons somewhere
in the middle which looks strange