Extract abstract superclass
I want to use the same config app for another widget, but that widget
only supports honeycomb or higher, so should not be referenced from
code that is used by lower android versions.
Therefore I cannot use *exactly* the same config app, but need two
versions.
So this commit extracts most code to a new abstract superclass, and
only the reference to the actual widget class is implemented in the
subclass.