Add widget for honeycomb and higher: using a listview
For older android versions that don't support ListView in
a widget the old widget with pages is kept.
I tested it with a simulator for android 2.3.1.  I verified that
indeed the old widget is shown, but I must say that not all
functionality worked there.  The 'next page' button didn't go
to the next page, and marking an item as finished didn't work
either.  I did the same tests with the master branch, and they
failed there as well, so I think my code didn't break anything
that was not already broken.