Extra functionality in widget
I think it is too easy to hit an item and mark it as 'bought' by
accident.  And after it happened, it is difficult to know which item
it was (if you even noticed that this happened).
Therefore I think it is better to display the 'bought' items grayed out.