Holo-style search / add in action bar.
The input panel and Add button on the bottom row are replaced by a Search
icon in the action bar. When it is activated, the action bar contents are
replaced by a search box. Text entered in the search box causes the list
to be filtered, and temporarily switched to Pick Items mode so that items
not currently on the list can still be found by the search. If search does
not find the item you can type its full name and then press enter or the
search widget's submit button and the item will be added to the list.

Still need to do something about capitalization preference, and ideally we
should change the icon of the search widget's submit button to indicate it
will add the item to the list.

The new behavior is only available from API 11 onwards. It can be disabled
by a new preference.