Commented out the AutoCompleteTextView for the search text field.
The completions were covering the list, which was unfortunate because the filtered list is much more functional than the autocomplete text view -- you can see whether the item you are searching for is already on the list, change its status or quantity, etc.

We could consider having a preference to choose the autocomplete popup menu instead, but if we do so, we would want to fix it first -- the items it was showing were not properly filtered.

Meanwhile if somebody absolutely must have the old way, there is a preference to turn off the "search/add in address bar" which will restore the old input area and button on the bottom of the screen.