Clear search field when drawer is opened.
Previously when opening the drawer, if the search field was expanded,
it would be collapsed but its contents would remain and we would still
be in search mode. This was confusing for users, who might for example
switch to a different list and wonder why fewer items were showing than
expected.

This fix may be a bit heavy-handed, but simpler attempts seem to be
thwarted by onRestoreInstanceState trying to put back the search
widget's text view's contents when the drawer is closed.