Apply filter values to added items.
In other words, if you have a tag filter active, and you go to
search mode and create a new item, or mark an existing item want
to buy, the tag from the filter will be added to the item created
or marked to buy.

Before this change, duplicate items might be created and still not
show on the screen after exiting search mode since they don't
match the filter.

Also, adding an item which already exists on the list no longer
marks it bought. That behavior was confusing.